chore: move `build_abi` logic to lib (#207)
- **chore: change `extract_abi_entries` arg type to
`&CompilationArtifact<DYLIB>`**
- **chore: move `extract_abi_entries` to `dylib` sub-module**
- **chore: change arg type of `CrateMetadata::resolve_output_dir` to
`camino::Utf8PathBuf`**
- **chore: move `CrateMetadata` to `cargo_near_build`**
- **chore: move types related to abi logic to lib**
- **chore: move `rite_to_file` to `cargo_near_build::near::abi`**
- **chore: move `BUILD_RS_ABI_STEP_HINT_ENV_KEY` to
`cargo_near_build::env_keys`**
- **chore: move `generate_abi` to
`cargo_near_build::near::abi::generate::procedure`**
- **chore: move `cargo_near::commands::abi_command::abi::run` to
`cargo_near_build::build_abi`**

---------
