+++ cargo-near-build/src/env_keys.rs
@@ -1,0 +1,1 @@
+pub const BUILD_RS_ABI_STEP_HINT_ENV_KEY: &str = "CARGO_NEAR_ABI_GENERATION";
