+++ cargo-near-build/src/lib.rs
@@ -2,6 +2,7 @@
 pub mod fs;
 // TODO: make mod non-pub
 pub mod cargo_native;
+pub mod env_keys;
 pub mod near;
 // TODO: consider making mod non-pub
 pub mod pretty_print;
@@ -10,3 +11,6 @@ pub mod types;
 
 // TODO: remove these exports, replace with `pub(crate)` visibility
 pub use cargo_native::{ArtifactType, DYLIB, WASM};
+
+pub use near::abi::build as build_abi;
+pub use types::near::abi::Opts as AbiOpts;
