+++ cargo-near-build/src/near/abi/generate/dylib.rs
@@ -3,10 +3,14 @@ use std::fs;
 
 use camino::Utf8Path;
 
+use crate::types::near::CompilationArtifact;
+use crate::DYLIB;
+
 // TODO: make func non-pub
 pub fn extract_abi_entries(
-    dylib_path: &Utf8Path,
+    artifact: &CompilationArtifact<DYLIB>,
 ) -> eyre::Result<Vec<near_abi::__private::ChunkedAbiEntry>> {
+    let dylib_path: &Utf8Path = &artifact.path;
     let dylib_file_contents = fs::read(dylib_path)?;
     let object = symbolic_debuginfo::Object::parse(&dylib_file_contents)?;
     log::debug!(
