+++ cargo-near-build/src/types/cargo/mod.rs
@@ -1,2 +1,4 @@
 // TODO: consider making mod non-pub
 pub mod manifest_path;
+// TODO: consider making mod non-pub
+pub mod metadata;
