+++ cargo-near/Cargo.toml
@@ -33,9 +33,7 @@ log = "0.4"
 rustc_version = "0.4"
 serde = "1.0.197"
 serde_json = "1.0"
-schemars = "0.8"
 near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
-zstd = "0.13"
 color-eyre = "0.6"
 inquire = "0.7"
 strum = { version = "0.24", features = ["derive"] }
