+++ cargo-near/src/types/mod.rs
@@ -1,4 +1,3 @@
 pub mod color_preference_cli;
-pub mod metadata;
 pub mod source_id;
 pub mod utf8_path_buf;
