+++ integration-tests/Cargo.toml
@@ -12,6 +12,7 @@ near-abi = "0.4.0"
 borsh = { version = "1.0.0", features = ["derive", "unstable__schema"] }
 camino = "1.1.1"
 cargo-near = { path = "../cargo-near" }
+cargo-near-build = { path = "../cargo-near-build" }
 color-eyre = "0.6"
 function_name = "0.3"
 git2 = "0.19"
